Создание отдельного презентера для поинтов
