+++ src/view/trip-edit-point-view.js
@@ -98,7 +98,7 @@ function createAddPointTemplate(point, allOffers, allDestination, pointDestinati
   </li>`
   );
 }
-export default class OffersView extends AbstractView{
+export default class EditorPointView extends AbstractView{
   #point = null;
   #allOffers = [];
   #allDestination = [];
